Removed old code... together with a few more fixes.
