+++ source/DasBlog.Web.UI/Views/Shared/_AdminNavBar.cshtml
@@ -2,7 +2,7 @@
 <div class="container" dasblog-authorized>
     <ul class="navbar-nav mr-auto">
         <li class="nav-item" ><a class="nav-link" asp-area="" asp-controller="post" asp-action="create">Create Post</a></li>
-        <li class="nav-item" ><a class="nav-link" asp-area="" asp-controller="users" asp-action="index">Users</a></li>
+        <li class="nav-item" ><a class="nav-link" asp-area="" asp-controller="UserAdmin" asp-action="index">Users</a></li>
         <li class="nav-item"><a class="nav-link" asp-area="" asp-controller="Activity" asp-action="ActivityList">Activity</a></li>
     </ul>
 </div>
