--- source/DasBlog.Web.UI/Views/Users/Components/UserList/Default.cshtml
@@ -1,61 +1,0 @@
-﻿@using System.Reflection.Metadata.Ecma335
-@using DasBlog.Core.Common
-@model System.Collections.Generic.IEnumerable<DasBlog.Core.Security.User>
-
-@functions
-{
-    string ElideTitle(string title)
-    {
-        const int maxLen = 20;
-        if (title.Length > maxLen)
-        {
-            return title.Substring(0, maxLen - 3) + "...";
-        }
-        else
-        {
-            return title;
-        }
-    }
-}
-
-
-<div class="row workaround-users-1 pull-left affix" style="background-colorx: green">
-    <div class="">
-        
-        <div class="row" style="background-colorx: red">
-            <div class="col-md-6 text-center h4" style="background-colorx: blue">Users</div>
-            <div class="col-md-6" style="background-colorx: yellow">
-                <div class="pull-right">
-                    <form asp-action="maintenance" asp-controller="users" method="get">
-                        @if (ViewBag.SubViewName == Constants.ViewUserSubView)
-                        {
-                            <input type="submit" class="btn btn-primary" name="maintenanceMode" value="@Constants.UsersCreateMode"/>
-                        }
-                        else
-                        {
-                            <input type="submit" class="btn" name="maintenanceMode" value="@Constants.UsersCreateMode" disabled/>
-
-                        }
-                    </form>
-                </div>
-
-            </div>
-        </div>
-    <div class="row workaround-users-2" >
-        <div class="col-md-12 users-sidebar">
-            <table class="table">
-                <tbody>
-                @foreach (var item in Model)
-                {
-                    <tr>
-                        <td>
-                            <user-link email-address="/users/@item.EmailAddress" linkability="@ViewBag.Linkability"> @ElideTitle(item.EmailAddress)</user-link>
-                        </td>
-                    </tr>
-                }
-                </tbody>
-            </table>
-        </div>
-    </div>
-    </div>
-</div>
