--- source/DasBlog.Web.UI/Views/Users/CreateUser.cshtml
@@ -1,16 +1,0 @@
-﻿@using DasBlog.Core.Common
-@model DasBlog.Web.Models.UsersViewModel
-
-<hr />
-<div class="row">
-    <div class="col-md-9">
-        <form asp-action="Maintenance">
-            <partial name="UserFields" model="Model" />
-            <div class="form-group">
-                <input type="submit" name="submitAction" value="@Constants.SaveAction" class="btn btn-default btn-primary" />
-                <input type="submit" name="submitAction" value="@Constants.CancelAction" class="btn cancel" />
-            </div>
-        </form>
-    </div>
-</div>
-
