--- source/DasBlog.Web.UI/Views/Users/DeleteUser.cshtml
@@ -1,16 +1,0 @@
-﻿@using DasBlog.Core.Common
-@model DasBlog.Web.Models.UsersViewModel
-
-<hr />
-<div class="row">
-    <div class="col-md-9">
-        <form asp-action="Maintenance">
-            <div class="form-group">
-                <span class="text-danger">Are you sure you wish to delete this user?</span>
-                <input type="submit" name="submitAction" value="@Constants.DeleteAction" class="btn btn-default btn-primary" />
-                <input type="submit" name="submitAction" value="@Constants.CancelAction" class="btn cancel" />
-            </div>
-            <partial name="UserFields" model="Model" />
-        </form>
-    </div>
-</div>
