--- source/DasBlog.Web.UI/Views/Users/EditUser.cshtml
@@ -1,17 +1,0 @@
-﻿@using DasBlog.Core.Common
-@using DasBlog.Core.Security
-@model DasBlog.Web.Models.UsersViewModel
-
-<hr />
-<div class="row">
-    <div class="col-md-9">
-        <form asp-action="Maintenance" name="editForm" id="editForm">
-            <partial name="UserFields" model="Model" />
-            <div class="form-group">
-                <input type="submit" name="submitAction" value="@Constants.SaveAction" class="btn btn-default btn-primary" />
-                <input type="submit" name="submitAction" value="@Constants.CancelAction" class="btn cancel" />
-            </div>
-        </form>
-    </div>
-</div>
-
