--- source/DasBlog.Web.UI/Views/Users/ErrorUser.cshtml
@@ -1,8 +1,0 @@
-﻿@model DasBlog.Web.Models.UsersViewModel
-
-<hr />
-<div class="row">
-</div>
-
-
-
