--- source/DasBlog.Web.UI/Views/Users/Maintenance.cshtml
@@ -1,32 +1,0 @@
-﻿@using DasBlog.Core.Common
-@model DasBlog.Web.Models.UsersViewModel
-@inject IDasBlogSettings dasBlogSettings
-
-@{
-    ViewBag.Title = "User Maintenance";
-}
-
-<form asp-action="Maintenance" asp-controller="Users" method="get">
-    @if (ViewBag.SubViewName == Constants.ViewUserSubView)
-    {
-        <input type="submit" class="btn btn-primary btn-default" name="maintenanceMode" value="@Constants.UsersEditMode" />
-        <input type="submit" class="btn" name="maintenanceMode" value="@Constants.UsersDeleteMode" />
-    }
-    else
-    {
-        <input type="submit" class="btn" name="maintenanceMode" value="@Constants.UsersEditMode" disabled />
-        <input type="submit" class="btn" name="maintenanceMode" value="@Constants.UsersDeleteMode" disabled />
-    }
-</form>
-
-<h2>@ViewBag.Title</h2>
-
-<partial name="@ViewBag.SubViewName" model="Model" />
-
-@*subview name is Edit or Delete or possibly something else*@
-
-@section Scripts {
-    <partial name="_ValidationScriptsPartial" />
-}
-
-
