+++ source/DasBlog.Web.UI/Views/Users/UserAdminFields.cshtml
@@ -2,8 +2,6 @@
 @using DasBlog.Core.Security
 @model DasBlog.Web.Models.UsersViewModel
 
-<div asp-validation-summary="ModelOnly" class="text-danger"></div>
-
 <input type="hidden" name="OriginalEmail" value="@Model.OriginalEmail" />        
 
 <div class="form-group row">
@@ -11,6 +9,12 @@
     <input name="EmailAddress" value="@Model.EmailAddress" type="email" class="form-control col-6" @ViewBag.Writability
            data-val-email="The Email field is not a valid e-mail address." data-val-required="The Email field is required."/>
     <span asp-validation-for="EmailAddress" class="text-danger"  data-valmsg-for="EmailAddress" data-valmsg-replace="true" ></span>
+</div>
+
+<div class="form-group row">
+    <label asp-for="DisplayName" class="col-form-label col-2"></label>
+    <input name="DisplayName" value="@Model.DisplayName" type="text" class="form-control col-6" @ViewBag.Writability />
+    <span asp-validation-for="DisplayName" class="text-danger" data-valmsg-for="DisplayName" data-valmsg-replace="true"></span>
 </div>
 
 <div class="form-group row">
@@ -25,45 +29,42 @@
         <select asp-for="Role" name="Role" asp-items="@Html.GetEnumSelectList<Role>()" type="text" class="form-control col-6" >
         </select>
     }
-    <span asp-validation-for="Role" class="text-danger"></span>
 </div>
+
 <div class="form-check row">
     <label asp-for="Ask" class="form-check-label col-2"></label>
-    <input name="Ask" checked="@Model.Ask" type="checkbox" class="form-check-input col-1" @ViewBag.Clickability />
-    <span asp-validation-for="Ask" class="text-danger"></span>
+    <input name="Ask" checked="@Model.Ask" type="checkbox" class="form-check-input col-1" @ViewBag.Clickability />    
 </div>
-<div class="form-group row">
-    <label asp-for="DisplayName" class="col-form-label col-2" ></label>
-    <input name="DisplayName" value="@Model.DisplayName" type="text" class="form-control col-6" @ViewBag.Writability />
-    <span asp-validation-for="DisplayName" class="text-danger"></span>
-</div>
+
 <div class="form-group row">
     <label asp-for="OpenIDUrl" class="col-form-label col-2" ></label>
-    <input name="OpenIDUrl" value="@Model.OpenIDUrl" type="text" class="form-control col-6" @ViewBag.Writability />
-    <span asp-validation-for="OpenIDUrl" class="text-danger"></span>
+    <input name="OpenIDUrl" value="@Model.OpenIDUrl" type="text" class="form-control col-6" @ViewBag.Writability />    
 </div>
+
 <div class="form-check row">
     <label asp-for="NotifyOnAllComment" class="form-check-label col-5"></label>
-    <input name="NotifyOnAllComment" checked="@Model.NotifyOnAllComment" type="checkbox" class="form-check-input col-1" @ViewBag.Clickability />
-    <span asp-validation-for="NotifyOnAllComment" class="text-danger"></span>
+    <input name="NotifyOnAllComment" checked="@Model.NotifyOnAllComment" type="checkbox" class="form-check-input col-1" @ViewBag.Clickability />    
 </div>
+
 <div class="form-check row">
     <label asp-for="NotifyOnOwnComment" class="form-check-label col-5" ></label>
     <input name="NotifyOnOwnComment" checked="@Model.NotifyOnOwnComment" type="checkbox" class="form-check-input col-1" @ViewBag.Clickability />
-    <span asp-validation-for="NotifyOnOwnComment" class="text-danger"></span>
 </div>
+
 <div class="form-check row">
     <label asp-for="NotifyOnNewPost" class="form-check-label col-5" ></label>
     <input name="NotifyOnNewPost" checked="@Model.NotifyOnNewPost" type="checkbox" class="form-check-input col-1" @ViewBag.Clickability />
-    <span asp-validation-for="NotifyOnNewPost" class="text-danger"></span>
 </div>
+
 <div class="form-check row">
     <label asp-for="Active" class="form-check-label col-5" ></label>
     <input name="Active" checked="@Model.Active" type="checkbox" class="form-check-input col-1" @ViewBag.Clickability />
-    <span asp-validation-for="Active" class="text-danger"></span>
 </div>
- <div class="form-group row">
-    <label asp-for="Password" class="col-form-label col-2" ></label>
+
+<div class="form-group row">
+    <label asp-for="Password" class="col-form-label col-2"></label>
     <input name="Password" value="@Model.Password" type="password" class="form-control col-6" @ViewBag.Writability />
-    <span asp-validation-for="Password" class="text-danger"></span>
+    <span asp-validation-for="Password" class="text-danger" data-valmsg-for="Password" data-valmsg-replace="true"></span>
 </div>
+
+ <div asp-validation-summary="ModelOnly" class="text-danger"></div>
