--- source/DasBlog.Web.UI/Views/Users/Users.cshtml
@@ -1,8 +1,0 @@
-﻿
-
-@{
-    ViewBag.Title = "title";
-    Layout = "_Layout";
-}
-
-<h2>title</h2>
