--- source/DasBlog.Web.UI/Views/Users/ViewUser.cshtml
@@ -1,11 +1,0 @@
-﻿@model DasBlog.Web.Models.UsersViewModel
-
-<hr />
-<div class="row">
-    <div class="col-md-9">
-        <form asp-action="Maintenance">
-            <partial name="UserFields" model="Model" />
-        </form>
-    </div>
-</div>
-
