--- source/DasBlog.Web.UI/Views/Users/_UsersLayout.cshtml
@@ -1,54 +1,0 @@
-﻿<html>
-<head>
-    <meta charset="utf-8"/>
-    <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
-    <title>@ViewData["Title"] - noter</title>
-    <environment include="Development">
-        <link rel="stylesheet" href="~/lib/bootstrap/dist/css/bootstrap.css"/>
-        <link href="~/lib/users/font-awesome.css" rel="stylesheet" />
-        <link href="~/lib/users/sb-admin-2.css" rel="stylesheet" />
-        <link rel="stylesheet" href="~/css/site.css"/>
-    </environment>
-    <environment exclude="Developent">
-        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
-        <link href="~/lib/users/font-awesome.css" rel="stylesheet" />
-        <link href="~/lib/users/sb-admin-2.css" rel="stylesheet" />
-        <link rel="stylesheet" href="~/css/site.css"/>
-    </environment>
-</head>
-<body>
-
-@{
-    ViewBag.Title = "title";
-}
-
-<partial name="_AdminNavBar" />
-
-<div class="workaround-users-3" role="navigation">
-    
-    <div class="pull-left affix" role="navigation">
-        @await Component.InvokeAsync("UserList")
-    </div>
-</div>
-
-<div id="page-wrapper">
-    @RenderBody()
-</div>
-
-<environment include="Development">
-    <script src="~/lib/jquery/dist/jquery.js"></script>
-    <script src="~/lib/jquery-validation/dist/jquery.validate.js"></script>
-    <script src="~/lib/jquery-validation-unobtrusive/jquery.validate.unobtrusive.js"></script>
-    <script src="~/lib/bootstrap/dist/js/bootstrap.js"></script>
-    <script src="~/js/site.js" asp-append-version="true"></script>
-</environment>
-<environment exclude="Development">
-    <script src="~/lib/jquery/dist/jquery.js"></script>
-    <script src="~/lib/jquery-validation/dist/jquery.validate.js"></script>
-    <script src="~/lib/jquery-validation-unobtrusive/jquery.validate.unobtrusive.js"></script>
-    <script src="~/lib/bootstrap/dist/js/bootstrap.js"></script>
-    <script src="~/js/site.js" asp-append-version="true"></script>
-</environment>
-@RenderSection("Scripts", required: false)
-</body>
-</html>
